Move `StoragePath` and `Transaction*` into `file-store-ot`.
The `Transaction*` classes aren't _strictly_ related to OT, but they're
close enough, and it makes better sense to have them there as (something like)
part of the lower-layer data structure definitions for file stuff.
