+++ doc/coding-conventions.md
@@ -94,7 +94,7 @@ taking into account recent additions to the language.
   import express from 'express';
   import fs from 'fs';
 
-  import { BaseFile, TransactionOp } from 'file-store';
+  import { BaseFile } from 'file-store';
   import { DataUtil, InfoError, Singleton } from 'util-common';
 
   import RegularBlort from './RegularBlort';
