+++ local-modules/doc-server/BaseControl.js
@@ -3,7 +3,8 @@
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
 import { Timeouts } from 'doc-common';
-import { Errors as fileStore_Errors, StoragePath, TransactionSpec } from 'file-store';
+import { Errors as fileStore_Errors } from 'file-store';
+import { StoragePath, TransactionSpec } from 'file-store-ot';
 import { BaseSnapshot, RevisionNumber } from 'ot-common';
 import { Delay } from 'promise-util';
 import { TBoolean, TFunction } from 'typecheck';
