+++ local-modules/doc-server/tests/test_BaseControl.js
@@ -10,8 +10,9 @@ import { Timeouts } from 'doc-common';
 import { MockChange, MockDelta, MockOp, MockSnapshot } from 'ot-common/mocks';
 import { DurableControl, FileAccess } from 'doc-server';
 import { MockControl } from 'doc-server/mocks';
-import { Errors as fileStore_Errors, TransactionSpec } from 'file-store';
+import { Errors as fileStore_Errors } from 'file-store';
 import { MockFile } from 'file-store/mocks';
+import { TransactionSpec } from 'file-store-ot';
 import { Timestamp } from 'ot-common';
 import { TheModule as mocks_TheModule } from 'ot-common/mocks';
 import { Errors, FrozenBuffer } from 'util-common';
