+++ local-modules/doc-server/tests/test_BaseDataManager.js
@@ -7,7 +7,7 @@ import { describe, it } from 'mocha';
 
 import { TheModule as appCommon_TheModule } from 'app-common';
 import { BaseDataManager, FileAccess, ValidationStatus } from 'doc-server';
-import { TransactionSpec } from 'file-store';
+import { TransactionSpec } from 'file-store-ot';
 import { MockFile } from 'file-store/mocks';
 
 /** {FileAccess} Convenient instance of `FileAccess`. */
