+++ local-modules/file-store-local/tests/test_LocalFile.js
@@ -6,8 +6,8 @@ import { assert } from 'chai';
 import fs from 'fs';
 import { after, describe, it } from 'mocha';
 
-import { TransactionOp, TransactionSpec } from 'file-store';
 import { LocalFile } from 'file-store-local';
+import { TransactionOp, TransactionSpec } from 'file-store-ot';
 import { FrozenBuffer } from 'util-common';
 
 import TempFiles from './TempFiles';
