+++ local-modules/file-store/Errors.js
@@ -2,10 +2,9 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
+import { StoragePath } from 'file-store-ot';
 import { TString } from 'typecheck';
 import { FrozenBuffer, InfoError, UtilityClass } from 'util-common';
-
-import StoragePath from './StoragePath';
 
 /**
  * Utility class for constructing errors salient to this module.
